requirements.txt
Django==1.11.16
django-braces==1.13.0
django-debug-toolbar==1.11
django-extensions==2.1.5
django-model-utils==3.1.2
django-star-ratings==0.7.0
djangorestframework==3.9.1
mysql-connector-python==8.0.11
mysqlclient==1.4.2
Pillow==5.3.0
protobuf==3.6.1
pytz==2018.9
six==1.12.0
sqlparse==0.2.4
swapper==1.1.0
